fix paths to components, rename components in footer
